requirements.txt
asgiref==3.8.1
Babel==2.16.0
colorama==0.4.6
ConfigParser==7.1.0
cryptography==44.0.0
cx_Oracle==8.3.0
docutils==0.21.2
email_validator==2.2.0
filelock==3.16.1
flask_babel==4.0.0
greenlet==3.1.1
HTMLParser==0.0.2
importlib_metadata==8.5.0
ipython==8.30.0
ipywidgets==8.1.5
jnius==1.1.0
keyring==25.5.0
Pillow==11.0.0
protobuf==5.29.1
pyOpenSSL==24.3.0
pysqlcipher3==1.2.0
pytest==8.3.4
python-dotenv==1.0.1
redis==5.2.1
Sphinx==8.1.3
sqlcipher3==0.5.4
thread==2.0.5
urllib3_secure_extra==0.1.0
watchdog==6.0.0
xmlrpclib==1.0.1
